-m "math_operations”

+++ functions/opdracht2/math_operations.py
@@ -5,16 +5,16 @@ def increment(nr2: float) -> float:
 def decrement(nr2: float) -> float:
   return nr2 - 1
   
-def add(nr1: float, nr2: float) -> int:
+def add(nr1: float, nr2: float) -> float:
   return nr1 + nr2 
 
-def substract(nr1: float, nr2: float) -> int:
+def substract(nr1: float, nr2: float) -> float:
   return nr1 - nr2
 
-def multiply(nr1: float, nr2: float) -> int:
+def multiply(nr1: float, nr2: float) -> float:
   return nr1 * nr2
 
-def divide(nr1: int, nr2: int) -> int:
+def divide(nr1: float, nr2: float) -> float:
   try:
     result = nr1 / nr2
     return result
